Divide Profile page into components
