+++ frontend/src/locales/ua/translation.json
@@ -80,7 +80,7 @@
   "upload-image": "Завантажити зображення",
   "username": "Ім'я користувача",
   "delete-button": "Видалити",
-  "e-mail": "E-mail:",
+  "e-mail": "Пошта:",
   "e-mail-input": "Ваша електронна пошта",
   "firstname": "Ім'я:",
   "firstname-input": "Ваше Ім'я",
