+++ frontend/src/styles/profile/deleteButton.css
@@ -1,0 +1,120 @@
+.delete-button {
+  position: absolute;
+  top: 20px;
+  right: 20px;
+
+  button {
+    --border: 5px; /* the border width */
+    --slant: 0.7em; /* control the slanted corners */
+    --color: #c51a00; /* the color */
+
+    width: 94px;
+    height: 50px;
+    border: none;
+    cursor: pointer;
+    font-weight: bold;
+    font-size: medium;
+    color: #fff;
+    background: linear-gradient(to bottom left, var(--color) 50%, #0000 50.1%)
+        top right,
+      linear-gradient(to top right, var(--color) 50%, #0000 50.1%) bottom left;
+    background-size: calc(var(--slant) + 1.3 * var(--border))
+      calc(var(--slant) + 1.3 * var(--border));
+    background-repeat: no-repeat;
+    box-shadow: 0 0 0 200px inset var(--s, #0000),
+      0 0 0 var(--border) inset var(--color);
+    clip-path: polygon(
+      0 0,
+      calc(100% - var(--slant)) 0,
+      100% var(--slant),
+      100% 100%,
+      var(--slant) 100%,
+      0 calc(100% - var(--slant))
+    );
+    transition: color var(--t, 0.3s), background-size 0.3s;
+  }
+  button:focus-visible {
+    outline-offset: calc(-1 * var(--border));
+    outline: var(--border) solid #000c;
+    clip-path: none;
+    background-size: 0 0;
+  }
+  button:hover,
+  button:active {
+    background-size: 100% 100%;
+    color: #fff;
+    --t: 0.2s 0.1s;
+  }
+  button:active {
+    --s: #0005;
+    transition: none;
+  }
+}
+
+/* Media Query For different screens */
+@media (min-width: 320px) and (max-width: 344px) {
+  .delete-button {
+    right: 4px;
+
+    button {
+      --border: 3px;
+      width: 58px;
+      height: 31px;
+      font-size: 10px;
+    }
+  }
+}
+@media (min-width: 345px) and (max-width: 430px) {
+  .delete-button button {
+    --border: 4px;
+    width: 62px;
+    height: 33px;
+    font-size: 10px;
+  }
+}
+@media (min-width: 345px) and (max-width: 389px) {
+  .delete-button {
+    right: 7px;
+  }
+}
+@media (min-width: 370px) and (max-width: 389px) {
+  .delete-button {
+    right: 6px;
+  }
+}
+@media (min-width: 390px) and (max-width: 399px) {
+  .delete-button {
+    right: 11px;
+  }
+}
+@media (min-width: 750px) and (max-width: 1200px) and (orientation: portrait) {
+  .delete-button button {
+    width: 98px;
+    height: 44px;
+  }
+}
+@media (min-width: 1300px) and (min-height: 912px) {
+  .delete-button button {
+    width: 98px;
+    height: 44px;
+  }
+}
+@media (min-width: 720px) and (max-height: 540px) {
+  .delete-button {
+    top: 14px;
+    right: 25px;
+
+    button {
+      width: 68px;
+      height: 34px;
+      font-size: 11px;
+    }
+  }
+}
+@media (min-width: 1024px) and (max-height: 600px) {
+  .delete-button button {
+    width: 80px;
+    height: 36px;
+    font-size: 13px;
+  }
+}
